+++ fabric/src/main/java/net/mehvahdjukaar/sleep_tight/mixins/fabric/GuiMixin.java
@@ -1,6 +1,7 @@
 package net.mehvahdjukaar.sleep_tight.mixins.fabric;
 
 import net.mehvahdjukaar.sleep_tight.SleepTightClient;
+import net.minecraft.client.DeltaTracker;
 import net.minecraft.client.gui.Gui;
 import net.minecraft.client.gui.GuiGraphics;
 import org.spongepowered.asm.mixin.Mixin;
@@ -12,7 +13,12 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 public class GuiMixin {
 
     @Inject(method = "renderExperienceBar", at = @At("HEAD"), cancellable = true)
-    public void supp$cannonCancelXPBar(GuiGraphics guiGraphics, int x, CallbackInfo ci) {
+    public void slepTight$bedCancelXpBar(GuiGraphics guiGraphics, int x, CallbackInfo ci) {
+        if (SleepTightClient.getLayingBedData() != null) ci.cancel();
+    }
+
+    @Inject(method = "renderExperienceLevel", at = @At("HEAD"), cancellable = true)
+    public void slepTight$bedCancelXpLevel(GuiGraphics guiGraphics, DeltaTracker deltaTracker, CallbackInfo ci) {
         if (SleepTightClient.getLayingBedData() != null) ci.cancel();
     }